Bug 932370 - Fix signed/unsigned comparison warning; r=bustage on a CLOSED TREE
UltraBlame original commit: f98f596ad05604ccbd44b26446cfec01b9b99078

+++ dom/events/ContentEventHandler.cpp
@@ -1010,7 +1010,7 @@ ContentEventHandler::GetFlatTextOffsetOfRange(nsIContent* aRootContent,
 
   nsCOMPtr<nsIContentIterator> iter = NS_NewContentIterator();
 
-  if (aNode->Length() >= aNodeOffset) {
+  if (aNode->Length() >= static_cast<uint32_t>(aNodeOffset)) {
     
     prev->SetEnd(startDOMNode, aNodeOffset);
     iter->Init(prev);
